pyproject.toml
[project]
name = "java-code-arena"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "carbon-api>=0.1",
    "flask>=3.1.2",
    "flask-cors>=6.0.2",
    "nanoid>=2.0.0",
    "pillow>=12.1.1",
    "waitress>=3.0.2",
]
